requirements.txt
asn1crypto==0.24.0
certifi==2017.11.5
cffi==1.11.3
chardet==3.0.4
cryptography==2.1.4
Cython==0.27.3
Django==2
elasticsearch==5.5.1
elasticsearch-dsl==5.4.0
ez-setup==0.9
idna==2.6
inflection==0.3.1
more-itertools==4.0.1
ndg-httpsclient==0.4.3
numpy==1.13.3
olefile==0.44
pandas==0.22.0
Pillow==4.3.0
Pillow-PIL==0.1.dev0
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
pytesseract==0.1.7
python-dateutil==2.6.1
pytz==2017.3
Quandl==3.3.0
requests==2.18.4
sh==1.12.14
six==1.11.0
urllib3==1.21.1
virtualenv==15.1.0
dj-database-url
django-excel
djangorestframework
ipdb
click
dj_database_url